+++ libs/aws/sabi-s3/src/core/verbs/to_endpoint.rs
@@ -1,14 +1,15 @@
-use crate::core::{S3Bucket, S3Result};
-use crate::verbs::HasObjectKey;
+use crate::core;
+use crate::core::verbs::HasObjectKey;
+use crate::core::S3Bucket;
 use reqwest::Url;
 use std::fmt::Debug;
 
 pub trait ToEndpoint: Debug {
-    fn to_endpoint(&self) -> S3Result<Url>;
+    fn to_endpoint(&self) -> core::Result<Url>;
 }
 
 impl<A: HasObjectKey> ToEndpoint for (&S3Bucket, &A) {
-    fn to_endpoint(&self) -> S3Result<Url> {
+    fn to_endpoint(&self) -> core::Result<Url> {
         let (bucket, request) = self;
         let full = format!(
             "https://{}.s3.amazonaws.com/{}",
