+++ libs/aws/sabi-s3/src/lib.rs
@@ -10,13 +10,3 @@ pub mod core;
 
 mod error;
 pub use error::Error;
-
-pub mod verbs;
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
