+++ libs/aws/sabi-s3/src/operations/put_object/file.rs
@@ -1,7 +1,7 @@
+use crate::core::verbs::HasObjectKey;
 use crate::internal;
 use crate::internal::{RequestResource, ResourceLoader};
 use crate::operations::put_object::RichFile;
-use crate::verbs::HasObjectKey;
 use reqwest::blocking::Body;
 use sabi_core::auth::v4::canonical::HashedPayload;
 use sabi_core::auth::v4::chrono::now;
